requirements.txt
Flask
requests
beautifulsoup4
openai>=1.0.0
cohere
python-dotenv
joblib
scikit-learn
